+++ src/elements/GeoList/GeoListGroup.vue
@@ -42,7 +42,7 @@ import cssSuffix from '../../mixins/cssModifierMixin'
 
 export default {
   name: 'GeoListGroup',
-  status: 'missing-tests',
+  status: 'ready',
   release: '4.0.0',
   mixins: [cssSuffix],
   props: {