+++ test/unit/specs/elements/GeoDropdown/GeoDropdownCompactButton.spec.js
@@ -1,36 +1,76 @@
+import _ from 'lodash'
 import { mount } from '@vue/test-utils'
 import GeoDropdownCompactButton from '@/elements/GeoDropdown/GeoDropdownCompactButton.vue'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { fas } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIconMock, expectFontAwesomeIconProp } from 'test/unit/utils/FontAwesomeIconMock'
 
 library.add(fas)
 
+const iconsToMock = [
+  'faEllipsisV'
+]
+const mockedFalIcons = _.mapValues(_.pick(fas, iconsToMock), (original) => {
+  return _.assign({}, original, {
+    prefix: 'fal'
+  })
+})
+library.add(mockedFalIcons)
+
 describe('GeoDropdownCompactButton', () => {
-  it('Should render icon', function () {
+  it('Should render GeoDropdwonCompactButton component', () => {
     const wrapper = mount(GeoDropdownCompactButton, {
       stubs: {
         'font-awesome-icon': FontAwesomeIcon
+      }
+    })
+    expect(wrapper.find('.geo-dropdown__compact-button-container').exists()).toBe(true)
+  })
+
+  it('Should render default icon', () => {
+    const wrapper = mount(GeoDropdownCompactButton, {
+      stubs: {
+        'font-awesome-icon': FontAwesomeIconMock
+      }
+    })
+    const fontAwesomeIconElem = wrapper.find(FontAwesomeIconMock)
+    expectFontAwesomeIconProp(fontAwesomeIconElem, ['fal', 'ellipsis-v'])
+  })
+
+  it('Should render correct icon when provided', () => {
+    const wrapper = mount(GeoDropdownCompactButton, {
+      stubs: {
+        'font-awesome-icon': FontAwesomeIconMock
       },
       propsData: {
         icon: ['fas', 'user']
       }
     })
-
-    expect(wrapper.find('svg[data-prefix="fas"][data-icon="user"]').exists()).toBe(true)
+    const fontAwesomeIconElem = wrapper.find(FontAwesomeIconMock)
+    expectFontAwesomeIconProp(fontAwesomeIconElem, ['fas', 'user'])
   })
 
-  it('Should emit click event when clicked', function () {
+  it('Should emit click event when clicked', () => {
     const wrapper = mount(GeoDropdownCompactButton, {
       stubs: {
         'font-awesome-icon': FontAwesomeIcon
-      },
-      propsData: {
-        icon: ['fas', 'user']
       }
     })
 
     wrapper.find('.geo-dropdown__compact-button-container').trigger('click')
     expect(wrapper.emitted()['click']).toBeTruthy()
+  })
+
+  it('Should apply a CSS suffix when the modifier is provided', function () {
+    const wrapper = mount(GeoDropdownCompactButton, {
+      propsData: {
+        cssModifier: 'test'
+      },
+      stubs: {
+        'font-awesome-icon': FontAwesomeIcon
+      }
+    })
+    expect(wrapper.find('.geo-dropdown__compact-button-container--test').exists()).toBe(true)
   })
 })